+++ flint/archive.py
@@ -183,7 +183,7 @@ def get_archive_options_from_yaml(strategy_yaml_path: Path) -> Dict[str, Any]:
         Dict[str, Any]: Loaded options for ArchiveOptions
     """
     archive_options = get_options_from_strategy(
-        strategy=strategy_yaml_path, mode="archive", round="initial"
+        strategy=strategy_yaml_path, mode="archive", round_info="initial"
     )
 
     logger.info(f"{archive_options=}")