+++ tests/test_configuration.py
@@ -1,6 +1,7 @@
 import filecmp
 from pathlib import Path
 
+from click import MissingParameter
 import pytest
 
 from flint.configuration import (
@@ -13,6 +14,7 @@ from flint.configuration import (
     get_selfcal_options_from_yaml,
     load_strategy_yaml,
     verify_configuration,
+    wrapper_options_from_strategy,
     write_strategy_to_yaml,
 )
 from flint.utils import get_packaged_resource_path
@@ -191,19 +193,19 @@ def test_get_options(strategy):
     # test to make sure we can generate a default strategy (see pytest fixture)
     # read it backinto a dict and then access some attributes
     wsclean = get_options_from_strategy(
-        strategy=strategy, mode="wsclean", round="initial"
+        strategy=strategy, mode="wsclean", round_info="initial"
     )
     assert isinstance(wsclean, dict)
     # example options
     assert wsclean["data_column"] == "CORRECTED_DATA"
 
-    wsclean = get_options_from_strategy(strategy=strategy, mode="wsclean", round=1)
+    wsclean = get_options_from_strategy(strategy=strategy, mode="wsclean", round_info=1)
     assert isinstance(wsclean, dict)
     # example options
     assert wsclean["data_column"] == "CORRECTED_DATA"
 
     archive = get_options_from_strategy(
-        strategy=strategy, mode="archive", round="initial"
+        strategy=strategy, mode="archive", round_info="initial"
     )
     assert isinstance(archive, dict)
     assert len(archive) > 0
@@ -216,14 +218,14 @@ def test_get_options(strategy):
 def test_get_mask_options(package_strategy):
     """Basic test to prove masking operation behaves well"""
     masking = get_options_from_strategy(
-        strategy=package_strategy, mode="masking", round="initial"
+        strategy=package_strategy, mode="masking", round_info="initial"
     )
 
     assert isinstance(masking, dict)
     assert masking["flood_fill_positive_seed_clip"] == 4.5
 
     masking2 = get_options_from_strategy(
-        strategy=package_strategy, mode="masking", round=1
+        strategy=package_strategy, mode="masking", round_info=1
     )
 
     print(strategy)
@@ -243,17 +245,15 @@ def test_get_mask_options_from_path(package_strategy_path):
     package_strategy = Path(package_strategy_path)
 
     masking = get_options_from_strategy(
-        strategy=package_strategy, mode="masking", round="initial"
+        strategy=package_strategy, mode="masking", round_info="initial"
     )
 
     assert isinstance(masking, dict)
     assert masking["flood_fill_positive_seed_clip"] == 4.5
 
     masking2 = get_options_from_strategy(
-        strategy=package_strategy, mode="masking", round=1
+        strategy=package_strategy, mode="masking", round_info=1
     )
-
-    print(strategy)
 
     assert isinstance(masking2, dict)
     assert masking2["flood_fill_positive_seed_clip"] == 40
@@ -265,13 +265,81 @@ def test_get_mask_options_from_path(package_strategy_path):
         assert masking[k] == masking2[k]
 
 
+@wrapper_options_from_strategy(update_options_keyword="update_options")
+def return_values(val, update_options=None):
+    "example doc string"
+    return val, update_options
+
+
+def test_wrapper_function_val(package_strategy_path):
+    """See if the wrapper to get strategy options from a file works nicely"""
+
+    # These asserts make sure that the  properties of the wrappede function are
+    # carried forward to the returned function
+    assert return_values.__name__ == "return_values"
+    assert return_values.__doc__ == "example doc string"
+    in_val = "ThisIsJack"
+    val, update_options = return_values(val=in_val, strategy=package_strategy_path)
+    assert val == in_val
+
+    # This makes sure that we can override the lookup from the strategy file
+    # by explicitly passing through the update_options keyword
+    val, update_options = return_values(
+        val=in_val, update_options="ignoring", strategy=package_strategy_path
+    )
+    assert val == in_val
+    assert update_options == "ignoring"
+
+    val, update_options = return_values(
+        val=in_val,
+        strategy=Path(package_strategy_path),
+        mode="masking",
+        round_info="initial",
+    )
+
+    assert val == in_val
+    assert isinstance(update_options, dict)
+    assert update_options["flood_fill_positive_seed_clip"] == 4.5
+
+
+@wrapper_options_from_strategy(update_options_keyword="update_options")
+def print_return_values(update_options):
+    "example doc string"
+    return "Jack", update_options
+
+
+def test_wrapper_function(package_strategy_path):
+    """See if the wrapper to get strategy options from a file works nicely"""
+
+    # These asserts make sure that the  properties of the wrappede function are
+    # carried forward to the returned function
+    assert print_return_values.__name__ == "print_return_values"
+    assert print_return_values.__doc__ == "example doc string"
+    val, update_options = print_return_values(strategy=package_strategy_path)
+    assert val == "Jack"
+
+    val, update_options = print_return_values(
+        strategy=Path(package_strategy_path), mode="masking", round_info="initial"
+    )
+
+    assert val == "Jack"
+    assert isinstance(update_options, dict)
+    assert update_options["flood_fill_positive_seed_clip"] == 4.5
+
+    with pytest.raises(MissingParameter):
+
+        @wrapper_options_from_strategy(update_options_keyword="JackNotHere")
+        def no_workie():
+            return 1
+
+
 def test_get_modes(package_strategy):
     # makes sure defaults for these modes are return when reuestion options
     # on a self-cal round without them set
     strategy = package_strategy
 
     for mode in ("wsclean", "gaincal", "masking"):
-        test = get_options_from_strategy(strategy=strategy, mode=mode, round=1)
+        test = get_options_from_strategy(strategy=strategy, mode=mode, round_info=1)
         assert isinstance(test, dict)
         assert len(test.keys()) > 0
 
@@ -279,10 +347,12 @@ def test_get_modes(package_strategy):
 def test_bad_round(package_strategy):
     # make sure incorrect round is handled properly
     with pytest.raises(AssertionError):
-        _ = get_options_from_strategy(strategy=package_strategy, round="doesnotexists")
+        _ = get_options_from_strategy(
+            strategy=package_strategy, round_info="doesnotexists"
+        )
 
     with pytest.raises(AssertionError):
-        _ = get_options_from_strategy(strategy=package_strategy, round=1.23456)
+        _ = get_options_from_strategy(strategy=package_strategy, round_info=1.23456)
 
 
 def test_empty_strategy_empty_options():
@@ -298,7 +368,7 @@ def test_max_round_override(package_strategy):
     # round is sound
     strategy = package_strategy
 
-    opts = get_options_from_strategy(strategy=strategy, round=9999)
+    opts = get_options_from_strategy(strategy=strategy, round_info=9999)
     assert isinstance(opts, dict)
     assert opts["data_column"] == "TheLastRoundIs3"
 
@@ -327,15 +397,19 @@ def test_updated_get_options(package_strategy):
     assert isinstance(strategy, Strategy)
 
     wsclean_init = get_options_from_strategy(
-        strategy=strategy, mode="wsclean", round="initial"
+        strategy=strategy, mode="wsclean", round_info="initial"
     )
     assert wsclean_init["data_column"] == "CORRECTED_DATA"
 
-    wsclean_1 = get_options_from_strategy(strategy=strategy, mode="wsclean", round=1)
+    wsclean_1 = get_options_from_strategy(
+        strategy=strategy, mode="wsclean", round_info=1
+    )
     assert wsclean_init["data_column"] == "CORRECTED_DATA"
     assert wsclean_1["data_column"] == "EXAMPLE"
 
-    wsclean_2 = get_options_from_strategy(strategy=strategy, mode="wsclean", round=2)
+    wsclean_2 = get_options_from_strategy(
+        strategy=strategy, mode="wsclean", round_info=2
+    )
     assert wsclean_2["multiscale"] is False
     assert wsclean_2["data_column"] == "CORRECTED_DATA"